% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_reference_levels}
\alias{add_reference_levels}
\title{Helper function for adding reference levels to the regression coefficients
of factors}
\usage{
add_reference_levels(coefsm, mod, or)
}
\arguments{
\item{coefsm}{a data frame containing the regression coefficients}

\item{mod}{the underlying model}

\item{or}{Whether the odds ratios should be displayed.}
}
\value{
a data frame containing the regression coefficients including their
reference levels
}
\description{
Helper function for adding reference levels to the regression coefficients
of factors
}
\keyword{internal}
