% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_rowlabs}
\alias{generate_rowlabs}
\title{Helper function for updating row names using the labels from a data frame}
\usage{
generate_rowlabs(
  rownames,
  mod,
  addref,
  rowlabs_auto,
  covar_names = names(stats::model.frame(mod))[-1],
  covar_pos
)
}
\arguments{
\item{rownames}{string vector with names from the object \code{coefsm}}

\item{mod}{A model}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{rowlabs_auto}{data frame used for automatically labeling rows using the
\code{"label"}-attributes of the columns}

\item{covar_names}{a string vector containing the names of the covariates
in the data set, default is \code{names(stats::model.frame(mod))[-1]}}

\item{covar_pos}{an integer vector containing the indexes of the rows
of \code{coefsm} that contain the covariates}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
Helper function for updating row names using the labels from a data frame
}
\keyword{internal}
