% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.coxph.R
\name{regtab.coxph}
\alias{regtab.coxph}
\title{Regression tables for coxph models}
\usage{
\method{regtab}{coxph}(
  mod,
  format = "latex",
  style_options = list(),
  hr = TRUE,
  loghr = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_loghr = FALSE,
  teststatistic = FALSE,
  pval = TRUE,
  caption = NULL,
  n_caption = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{coxph}.}

\item{format}{Type of output corresponding to the \code{format} argument
in \code{kbl::kableExtra}. For LaTeX output use \code{"latex"},
for HTML output use \code{"html"} and for Word output use
\code{"pipe"} or \code{"simple"}.}

\item{style_options}{A list of further options for modyfing the output,
corresponding to the arguments of \code{kableExtra::kable_styling}.}

\item{hr}{Whether the hazard ratios should be displayed if
\code{results = "summary"}.}

\item{loghr}{Whether the log hazard ratios should be displayed.}

\item{ci}{Whether confidence intervals for the hazard ratios should be
displayed if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_loghr}{Whether the standard errors of the log harzard ratios
should be displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{caption}{Title that is used as the caption of the LaTeX table.
If \code{title = NULL} then the caption is ”Cox Regression”.}

\item{n_caption}{Wheter the sample size that was used to compute the model
should be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in table.
If rowlabs is NULL then the rownames that are shown in the output
of \code{summary} are used.}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{digits}{Number of digits that are displayed in the tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
regtab method for models of class \code{coxph}.
}
\examples{
lung <- survival::lung
lung.coxph <- survival::coxph(survival::Surv(time, status) ~
  age + factor(sex) + meal.cal + ph.karno, data = lung)
regtab(lung.coxph)
}
