% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.glm.R
\name{regtab.glm}
\alias{regtab.glm}
\title{Regression tables for glm models}
\usage{
\method{regtab}{glm}(
  mod,
  or = TRUE,
  logor = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_logor = FALSE,
  vcov = NULL,
  teststatistic = FALSE,
  pval = TRUE,
  intercept = FALSE,
  caption = NULL,
  n_caption = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{glm}.}

\item{or}{Whether the odds ratios should be displayed.}

\item{logor}{Whether the log odds ratios should be displayed.}

\item{ci}{Whether confidence intervals for the odds ratios should be
displayed.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_logor}{Whether the standard errors of the log odds ratios
should be displayed if \code{results = "summary"}.}

\item{vcov}{If \code{NULL} then the covariance matrix computed
from \code{lm} will be used. Alternatively, one of the covariance matrix
functions from the the \code{sandwich} package can be passed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be
displayed.}

\item{caption}{Caption is used for the table. If \code{caption = NULL}
then the caption is ”Logistic Regression”.}

\item{n_caption}{Wheter the sample size that was used to compute the model
should be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in table.
If rowlabs is NULL then the rownames that are shown in the output
of \code{summary} are used.}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{digits}{Number of digits that are displayed in the tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
regtab method for models of class \code{glm}.
}
\details{
Models of class \code{glm} are currently only supported for
logistic regression model.
}
\examples{
iris$Sepal.Length <- ifelse(iris$Sepal.Length < median(iris$Sepal.Length), 0, 1)
iris.glm <- glm(Sepal.Length ~ Sepal.Width + Petal.Width + Species,
  data = iris,
  family = binomial
  )
regtab(iris.glm)
}
