% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.lm.R
\name{regtab.lm}
\alias{regtab.lm}
\title{Regression tables for lm models}
\usage{
\method{regtab}{lm}(
  mod,
  format = "latex",
  style_options = list(),
  ci = TRUE,
  ci_level = 0.95,
  se = FALSE,
  vcov = NULL,
  teststatistic = FALSE,
  pval = TRUE,
  intercept = FALSE,
  caption = NULL,
  n_caption = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lm}.}

\item{format}{Type of output corresponding to the \code{format} argument
in \code{kbl::kableExtra}. For LaTeX output use \code{"latex"},
for HTML output use \code{"html"} and for Word output use
\code{"pipe"} or \code{"simple"}.}

\item{style_options}{A list of further options for modyfing the output,
corresponding to the arguments of \code{kableExtra::kable_styling}.}

\item{ci}{Whether confidence intervals for the estimates should be displayed.}

\item{ci_level}{Level for the confidence intervals.}

\item{se}{Whether the standard errors of the estimates should be
displayed.}

\item{vcov}{If \code{NULL} then the covariance matrix computed
from \code{lm} will be used. Alternatively, one of the covariance matrix
functions from the the \code{sandwich} package can be passed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be
displayed.}

\item{caption}{Caption is used for the table. If \code{caption = NULL}
then the caption is ”Linear Regression”.}

\item{n_caption}{Whether the sample size that was used to compute the model
should be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in table.
If rowlabs is NULL then the rownames that are shown in the output
of \code{summary} are used.}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{digits}{Number of digits that are displayed in the tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
regtab method for models of class \code{lm}.
}
\examples{
iris.lm <- lm(Sepal.Length ~ Sepal.Width + Petal.Width + Species, data = iris)
regtab(iris.lm)
}
