% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.lme.R
\name{regtab.lme}
\alias{regtab.lme}
\title{Regression tables for lme models}
\usage{
\method{regtab}{lme}(
  mod,
  format = "latex",
  style_options = list(),
  ci = TRUE,
  ci_level = 0.95,
  se = FALSE,
  teststatistic = FALSE,
  df = TRUE,
  pval = TRUE,
  intercept = FALSE,
  caption = NULL,
  n_caption = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lme}.}

\item{format}{Type of output corresponding to the \code{format} argument
in \code{kbl::kableExtra}. For LaTeX output use \code{"latex"},
for HTML output use \code{"html"} and for Word output use
\code{"pipe"} or \code{"simple"}.}

\item{style_options}{A list of further options for modyfing the output,
corresponding to the arguments of \code{kableExtra::kable_styling}.}

\item{ci}{Whether confidence intervals for the estimates should be displayed.}

\item{ci_level}{Level for the confidence intervals.}

\item{se}{Whether the standard errors of the estimates should be
displayed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if
\code{results = "Anova"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be
displayed.}

\item{caption}{Caption that is used for the table. If \code{caption = NULL}
then the caption is "Mixed Model Regression".}

\item{n_caption}{Whether the sample size that was used to compute the model
should be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in table.
If rowlabs is NULL then the rownames that are shown in the output
of \code{summary} are used.}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{digits}{Number of digits that are displayed in the tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
regtab method for models of class \code{lme}.
}
\examples{
library(nlme)
bdf.lme <- nlme::lme(IQ.verb ~ sex + aritPOST + denomina + Minority,
  random = ~1|schoolNR,
  data = nlme::bdf)
regtab(bdf.lme)
}
