% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.polr.R
\name{regtab.polr}
\alias{regtab.polr}
\title{Regression tables for proportional odds regressions models from the MASS
package}
\usage{
\method{regtab}{polr}(
  mod,
  format = "latex",
  style_options = list(),
  or = TRUE,
  logor = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_logor = FALSE,
  vcov = NULL,
  teststatistic = FALSE,
  pval = TRUE,
  intercept = FALSE,
  caption = NULL,
  n_caption = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  rowlabs_auto = NULL,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{polr}.}

\item{format}{Type of output corresponding to the \code{format} argument
in \code{kbl::kableExtra}. For LaTeX output use \code{"latex"},
for HTML output use \code{"html"} and for Word output use
\code{"pipe"} or \code{"simple"}.}

\item{style_options}{A list of further options for modyfing the output,
corresponding to the arguments of \code{kableExtra::kable_styling}.}

\item{or}{Whether the odds ratios should be displayed.}

\item{logor}{Whether the log odds ratios should be displayed.}

\item{ci}{Whether confidence intervals for the odds ratios should be
displayed.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_logor}{Whether the standard errors of the log odds ratios
should be displayed.}

\item{vcov}{If \code{NULL} then the covariance matrix computed
from \code{lm} will be used. Alternatively, one of the covariance matrix
functions from the the \code{sandwich} package can be passed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be
displayed.}

\item{caption}{Caption that is used for the table. If \code{caption = NULL}
then the caption is ”Logistic Regression”.}

\item{n_caption}{Whether the sample size that was used to compute the model
should be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in table.
If rowlabs is NULL then the rownames that are shown in the output
of \code{summary} are used.}

\item{addref}{Whether a row for the reference category of categorial
variables should be displayed.}

\item{digits}{Number of digits that are displayed in the tables.}

\item{rowlabs_auto}{data frame used for automatically labeling rows using the
\code{"label"}-attributes of the columns}

\item{...}{Further optional arguments.}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
regtab method for models of class \code{polr}.
}
\details{
Models of class \code{polr} are currently only supported for
logistic method. If \code{pval = TRUE}, then the function \code{AER:::coeftest.polr}
is used for for calculating p-values. Note that calculating p-values
for proportional odds models may be unreliable.
}
\examples{
library(MASS)
# Basic example taken from the MASS documentation
data(housing, package="MASS")
attr(housing$Sat, "label") <- "Satisfaction"
attr(housing$Infl, "label") <- "Degree of influence"
attr(housing$Type, "label") <- "Type"
attr(housing$Cont, "label") <- "Contact with other residents"
attr(housing$Freq, "label") <- "Number of residents"
house.plr <- MASS::polr(Sat ~ Infl + Type + Cont, weights = Freq,
                        Hess = TRUE, data = housing)
regtab(house.plr, rowlabs_auto = housing)
}
