% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{regtab}
\alias{regtab}
\title{Regression Tables}
\usage{
regtab(mod, ...)
}
\arguments{
\item{mod}{A model object for which a table should created.}

\item{...}{Further optional arguments.}
}
\value{
\code{regtab} uses \code{kableExtra::kbl} to return a table.
}
\description{
regtab is a generic function that creates tables for the results of
various regression models.
}
